WA-284 Removed files
