--- Ethereum/Ethereum/Ethereum.swift
@@ -1,7 +1,0 @@
-//
-//  Copyright © 2018 Gnosis. All rights reserved.
-//
-
-class Ethereum {
-
-}
