+++ IdentityAccess/IdentityAccessDomainModel/AccountTests.swift
@@ -320,111 +320,5 @@ extension AccountTests {
 
 }
 
-class InMemoryUserDefaults: KeyValueStore {
 
-    var dict = [String: Any]()
 
-    func bool(for key: String) -> Bool? {
-        return dict[key] as? Bool
-    }
-
-    func setBool(_ value: Bool, for key: String) {
-        dict[key] = value
-    }
-
-    func int(for key: String) -> Int? {
-        return dict[key] as? Int
-    }
-
-    func setInt(_ value: Int, for key: String) {
-        dict[key] = value
-    }
-
-    func deleteKey(_ key: String) {
-        dict.removeValue(forKey: key)
-    }
-
-}
-
-class MockKeychain: SecureStore {
-
-    private var storedPassword: String?
-    var throwsOnSavePassword = false
-    var throwsOnGetPassword = false
-    var throwsOnRemovePassword = false
-
-    enum Error: Swift.Error {
-        case error
-    }
-
-    func password() throws -> String? {
-        if throwsOnGetPassword {
-            throw MockKeychain.Error.error
-        }
-        return storedPassword
-    }
-
-    func savePassword(_ password: String) throws {
-        if throwsOnSavePassword {
-            throw MockKeychain.Error.error
-        }
-        storedPassword = password
-    }
-
-    func removePassword() throws {
-        if throwsOnRemovePassword {
-            throw MockKeychain.Error.error
-        }
-        storedPassword = nil
-    }
-
-    func privateKey() throws -> PrivateKey? { return nil }
-
-    func savePrivateKey(_ privateKey: PrivateKey) throws {}
-
-    func removePrivateKey() throws {}
-
-    func mnemonic() throws -> Mnemonic? { return nil }
-
-    func saveMnemonic(_ mnemonic: Mnemonic) throws {}
-
-    func removeMnemonic() throws {}
-
-}
-
-class MockBiometricService: BiometricAuthenticationService {
-
-    var biometryType: BiometryType = .none
-    private var savedActivationCompletion: (() -> Void)?
-    var shouldActivateImmediately = false
-    var biometryAuthenticationResult = true
-    var isAuthenticationAvailable = false
-
-    private var savedAuthenticationCompletion: ((Bool) -> Void)?
-    var shouldAuthenticateImmediately = false
-
-    func activate(completion: @escaping () -> Void) {
-        if shouldActivateImmediately {
-            completion()
-        } else {
-            savedActivationCompletion = completion
-        }
-    }
-
-    func completeActivation() {
-        savedActivationCompletion?()
-    }
-
-    func authenticate(completion: @escaping (Bool) -> Void) {
-        if shouldAuthenticateImmediately {
-            completion(biometryAuthenticationResult)
-        } else {
-            savedAuthenticationCompletion = completion
-        }
-    }
-
-    func completeAuthentication(result: Bool) {
-        savedAuthenticationCompletion?(result)
-    }
-
-}
