--- MultiSigWallet/MultiSigWallet/MultiSigWalletApplication/MultiSigWalletApplication/MultiSigWalletApplication.swift
@@ -1,7 +1,0 @@
-//
-//  Copyright © 2018 Gnosis. All rights reserved.
-//
-
-class MultiSigWalletApplication {
-
-}
