--- MultiSigWallet/MultiSigWallet/MultiSigWalletPortAdapter/MultiSigWalletPortAdapter/MultiSigWalletPortAdapter.swift
@@ -1,7 +1,0 @@
-//
-//  Copyright © 2018 Gnosis. All rights reserved.
-//
-
-class MultiSigWalletPortAdapter {
-
-}
